requirements.txt
streamlit
pdf2image
pytesseract
pandas 
numpy
regex
bs4
requests
spacy
threadpoolctl
transformers
sklearn
nltk
matplotlib
wordcloud
gensim.models
seaborn
streamlit_option_menu
base64
sentencepiece
